+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file ordered by n
 
 
 ## [1.0.1] - 2023-10-04 - Bruno Gritti
+ 
+#### Added
+ - Factories and command to generate test data
+#### Changed
+#### Fixed
+
+---
+
+## [1.0.1] - 2023-10-04 - Bruno Gritti
   
 First deploy - migrations and seeders